svg/interact/scripted/module-01.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><title>script type=&quot;module&quot;</title><metadata><h:link href="https://svgwg.org/svg2-draft/interact.html#ScriptElement" rel="help"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><script type="module">window.evaluated_module_script = true;</script><script>var test = async_test(&quot;module script in SVG documents should be evaluated.&quot;); window.addEventListener(&quot;load&quot;, () =&gt; { test.step(() =&gt; { assert_true(window.evaluated_module_script); }); test.done(); });</script></svg>
